Move reporting hostname to analytics.py
There was a bug with the ordering of settings files which caused that
reported metrics get reported with 'localhost' domain.

We constructed the MACHINE_HOSTNAME in the settings_common.py file,
which was problematic if we updated HOSTNAME in settings.prod.py or some
other file later.

Moving the preparation of the hostname to the anaytics.py fixes this
issue.