Cache generator as a list to determine it's size
Fixes Sentry: https://sentry.io/nyaruka/rapidpro/issues/794129979

Added a constant THRESHOLD_LOG_INFO_DURING_EXPORT to ExportTask models
for Contacts, FlowRun and Msgs. This constant is used to determine when
we log a informational message about the current state of the export.

By adding this constant we can modify its value in tests and actually
test that we are correctly logging export task state.