Don't overwrite `show_in_table` flag on contact import
This PR fixes prevents overwriting of `show_in_table` flag for existing
fields when importing contacts.

Also fixes an edge case behavior when importing contacts that have
`label` string in the contact field label. This was caused by a regex
that was used to parse posted data which should have checked the end of
the string. It only occurred when more then one field had label with
'label' in the name.